Get blocks array on page
