+++ lib/block.js
@@ -8,9 +8,10 @@ function Block(options, ctx, canvas) {
   this.canvas = canvas;
 }
 
-Block.prototype.draw = function() {
-  this.ctx.fillRect(this.x, this.y, this.width, this.height);
-};
+//this is no longer doing anything//
+  // Block.prototype.draw = function() {
+  //   this.ctx.fillRect(this.x, this.y, this.width, this.height);
+  // };
