+++ lib/game.js
@@ -15,23 +15,9 @@ function Game(ctx, canvas) {
 Game.prototype.run = function() {
   this.paddle.draw(this.ctx);
   this.ball.draw();
-  this.createBlocks();
   this.ballMovement();
   this.ballCollisionWithWall();
   this.ballCollisionWithPaddle();
-};
-
-Game.prototype.createBlocks = function() {
-  var blocksArray = [];
-  for(i = 0; i < 8; i++){
-    for(j = 0; j < 4; j++) {
-    blocksArray.push(new Block({ x:2 + i * 100, y:50 + j * 25, width:96, height:20 } ));
-    }
-  }
-  debugger;
-  blocksArray.forEach(function(block) {
-  this.block.draw();
-  });
 };
 
 Game.prototype.ballMovement = function() {
